// src/http-busca-search/models/githubComPrefeituraRioAppBuscaSearchInternalModelsFieldChange.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */

export interface GithubComPrefeituraRioAppBuscaSearchInternalModelsFieldChange {
  field_name?: string
  new_value?: unknown
  old_value?: unknown
  /** "string", "int", "bool", "array", "object" */
  value_type?: string
}