// src/http-busca-search/models/githubComPrefeituraRioAppBuscaSearchInternalModelsVersionDiff.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */
import type { GithubComPrefeituraRioAppBuscaSearchInternalModelsFieldChange } from './githubComPrefeituraRioAppBuscaSearchInternalModelsFieldChange'

export interface GithubComPrefeituraRioAppBuscaSearchInternalModelsVersionDiff {
  change_type?: string
  changed_at?: number
  changed_by?: string
  changes?: GithubComPrefeituraRioAppBuscaSearchInternalModelsFieldChange[]
  from_version?: number
  to_version?: number
}